// frontend/src/backend.d.ts
export type Time = bigint;
export interface InstagramFeedItem {
    id: string;
    link?: string;
    published: boolean;
    mediaUrl: string;
    story: boolean;
    timestamp: Time;
    caption: string;
    mediaType: string;
}